refactor: 시장 페이지 관련 기능 리팩토링
