+++ src/main/java/com/themore/moamoatown/job/service/JobService.java
@@ -17,7 +17,6 @@ import java.util.List;
  * 2024.08.26   임재성        역할 리스트 조회 기능 추가
  * 2024.08.26   임재성        역할 리스트 조회 메서드 수정
  * 2024.08.26   임재성        역할 요청 기능 추가
- * 2024.08.26   임원정        getJobRequests, createJob, allowJobRequest 추가
  * </pre>
  */
 public interface JobService {
@@ -35,11 +34,4 @@ public interface JobService {
      * @return 역할 요청 처리 결과
      */
     JobApplyResponseDTO requestJob(JobRequestDTO jobRequestDTO);
-
-    // 역할 요청 현황 조회
-    List<JobRequestsResponseDTO> getJobRequests(Long townId);
-    // 역할 만들기
-    void createJob(JobCreateRequestDTO requestDTO, Long townId);
-    // 역할 선정
-    void allowJobRequest(Long jobRequestId);
 }
