+++ src/main/java/com/themore/moamoatown/job/service/JobServiceImpl.java
@@ -1,18 +1,18 @@
 package com.themore.moamoatown.job.service;
 
 import com.themore.moamoatown.common.exception.CustomException;
-import com.themore.moamoatown.job.dto.*;
+import com.themore.moamoatown.job.dto.JobApplyResponseDTO;
+import com.themore.moamoatown.job.dto.JobRequestDTO;
+import com.themore.moamoatown.job.dto.JobResponseDTO;
 import com.themore.moamoatown.job.mapper.JobMapper;
-import com.themore.moamoatown.notification.service.NotificationService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
-import static com.themore.moamoatown.common.exception.ErrorCode.*;
+import static com.themore.moamoatown.common.exception.ErrorCode.JOB_APPLY_INSERT_FAILED;
 
 
 /**
@@ -29,8 +29,6 @@ import static com.themore.moamoatown.common.exception.ErrorCode.*;
  * 2024.08.26   임재성        역할 리스트 조회 기능 추가
  * 2024.08.26   임재성        역할 리스트 조회 메서드 수정
  * 2024.08.26   임재성        역할 요청 기능 추가
- * 2024.08.26   임원정        getJobRequests, createJob, allowJobRequest 추가
- * 2024.08.28  임원정         알림 전송 로직 추가
  * </pre>
  */
 @Log4j
@@ -39,7 +37,6 @@ import static com.themore.moamoatown.common.exception.ErrorCode.*;
 public class JobServiceImpl implements JobService {
 
     private final JobMapper jobMapper;
-    private final NotificationService notificationService;
     /**
      * 타운 ID를 통해 JOB 리스트를 조회합니다.
      * @param townId 타운 ID
@@ -69,58 +66,5 @@ public class JobServiceImpl implements JobService {
         return JobApplyResponseDTO.builder()
                 .message("역할 요청이 성공적으로 처리되었습니다.")
                 .build();
-    }
-
-    /**
-     * 타운 내 역할 신청 현황 조회
-     * @param townId
-     * @return
-     */
-    @Override
-    @Transactional
-    public List<JobRequestsResponseDTO> getJobRequests(Long townId){
-        return jobMapper.selectJobRequestByTownId(townId)
-                .stream()
-                .map(jobRequest -> JobRequestsResponseDTO.builder()
-                        .jobRequestId(jobRequest.getJobRequestId())
-                        .name(jobRequest.getName())
-                        .comments(jobRequest.getComments())
-                        .nickName(jobRequest.getNickName())
-                        .allowYN(jobRequest.getAllowYN())
-                        .build())
-                .collect(Collectors.toList());
-    }
-
-    /**
-     * 역할 생성
-     * @param requestDTO
-     * @param townId
-     */
-    @Override
-    @Transactional
-    public void createJob (JobCreateRequestDTO requestDTO, Long townId){
-        JobCreateRequestDTO jobCreateRequestDTO = JobCreateRequestDTO.builder()
-                .name(requestDTO.getName())
-                .description(requestDTO.getDescription())
-                .pay(requestDTO.getPay())
-                .townId(townId)
-                .build();
-        if(jobMapper.insertJob(jobCreateRequestDTO) != 1) throw new CustomException(JOB_CREATE_FAILED);
-    }
-
-    /**
-     * 역할 선정
-     * @param jobRequestId
-     */
-    @Override
-    @Transactional
-    public void allowJobRequest(Long jobRequestId) {
-        if(jobMapper.updateJobRequestAllowed(jobRequestId) != 1) throw new CustomException(JOB_REQUEST_ALLOW_FAILED);
-        // 역할이 승인된 회원 정보 조회
-        Long memberId = jobMapper.findMemberIdByJobRequestId(jobRequestId);
-
-        // 알림 전송
-        String content = "축하합니다! 요청하신 역할에 선정되었습니다.";
-        notificationService.notifyMember(memberId, content, "job");
     }
 }
