+++ src/main/java/com/themore/moamoatown/quest/mapper/QuestMapper.java
@@ -1,9 +1,6 @@
 package com.themore.moamoatown.quest.mapper;
 
-import com.themore.moamoatown.quest.dto.MemberQuestRequestsResponseDTO;
-import com.themore.moamoatown.quest.dto.QuestCreateRequestDTO;
 import com.themore.moamoatown.quest.dto.QuestResponseDTO;
-import com.themore.moamoatown.quest.dto.QuestStatusListResponseDTO;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
@@ -20,9 +17,6 @@ import java.util.List;
  * ----------  --------    ---------------------------
  * 2024.08.26  이주현        최초 생성
  * 2024.08.26  이주현        퀘스트 수락 요청 기능 추가
- * 2024.08.27  임원정        insertQuest, selectQuestStatusListByTownId 메소드 추가
- * 2024.08.28  임원정        selectMemberQuestByQuestId, updateMemberQuestSelected 메소드 추가
- * 2024.08.28  임원정        callCompleteQuestProcedure, findMemberIdByMemberQuestId 메소드 추가
  * </pre>
  */
 
@@ -33,16 +27,4 @@ public interface QuestMapper {
 
     // 퀘스트 수락 요청
     int insertMemberQuest(@Param("memberId") Long memberId, @Param("questId") Long questId);
-    // 퀘스트 생성
-    int insertQuest(QuestCreateRequestDTO questCreateRequestDTO);
-    // 퀘스트 현황 리스트 조회
-    List<QuestStatusListResponseDTO> selectQuestStatusListByTownId(Long townId);
-    // 퀘스트 신청 조회
-    List<MemberQuestRequestsResponseDTO> selectMemberQuestByQuestId(Long questId);
-    // 퀘스트 수행인 선정
-    int updateMemberQuestSelected(Long memberQuestId);
-    // 퀘스트 완료 처리
-    void callCompleteQuestProcedure(Long memberQuestId);
-    // 퀘스트 요청 또는 완료 시 회원 ID 조회
-    Long findMemberIdByMemberQuestId(Long memberQuestId);
 }
