+++ src/main/java/com/themore/moamoatown/quest/service/QuestService.java
@@ -1,9 +1,6 @@
 package com.themore.moamoatown.quest.service;
 
-import com.themore.moamoatown.quest.dto.MemberQuestRequestsResponseDTO;
-import com.themore.moamoatown.quest.dto.QuestCreateRequestDTO;
 import com.themore.moamoatown.quest.dto.QuestResponseDTO;
-import com.themore.moamoatown.quest.dto.QuestStatusListResponseDTO;
 
 import java.util.List;
 
@@ -18,9 +15,6 @@ import java.util.List;
  * ----------  --------    ---------------------------
  * 2024.08.26  이주현        최초 생성
  * 2024.08.26  이주현        퀘스트 수락 요청 기능 추가
- * 2024.08.27  임원정        퀘스트 생성, 퀘스트 현황 리스트 추가
- * 2024.08.28  임원정        퀘스트 요청 조회, 퀘스트 수행인 선정 추가
- * 2024.08.28  임원정        퀘스트 요청 완료 처리 추가
  * </pre>
  */
 
@@ -30,14 +24,4 @@ public interface QuestService {
 
     // 퀘스트 수락 요청
     void addMemberQuest(Long memberId, Long questId);
-    // 퀘스트 생성
-    void createQuest(QuestCreateRequestDTO requestDTO, Long townId);
-    // 퀘스트 현황 리스트 조회
-    List<QuestStatusListResponseDTO> getQuestStatusList(Long townId);
-    // 퀘스트 요청(memberQuest) 조회
-    List<MemberQuestRequestsResponseDTO> getMemberQuests(Long questId);
-    // 퀘스트 수행인 선정
-    void updateMemberQuestSelected(Long memberQuestId);
-    // 퀘스트 요청 완료 처리
-    void completeMemberQuest(Long memberQuestId);
 }
