+++ src/main/java/com/themore/moamoatown/town/dto/QuestCreateRequestDTO.java
@@ -1,12 +1,14 @@
-package com.themore.moamoatown.quest.dto;
+package com.themore.moamoatown.town.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import javax.validation.constraints.NotNull;
+
 /**
- * 퀘스트 현황 조회 Response DTO
+ * 퀘스트 생성 Request DTO
  * @author 임원정
  * @since 2024.08.27
  * @version 1.0
@@ -22,12 +24,16 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class QuestStatusListResponseDTO {
-    private Long questId;
+public class QuestCreateRequestDTO {
+    @NotNull
     private String title;
+    @NotNull
+    private String description;
+    @NotNull
     private Long reward;
-    private String deadline;
-    private Long requestCnt;
-    private Long selectedCnt;
+    @NotNull
     private Long capacity;
+    @NotNull
+    private String deadline;
+    private Long townId;
 }
