+++ src/main/java/com/themore/moamoatown/wish/controller/WishController.java
@@ -3,14 +3,15 @@ package com.themore.moamoatown.wish.controller;
 import com.themore.moamoatown.common.annotation.Auth;
 import com.themore.moamoatown.common.annotation.MemberId;
 import com.themore.moamoatown.common.annotation.TownId;
-import com.themore.moamoatown.wish.dto.*;
+import com.themore.moamoatown.wish.dto.WishItemPurchaseRequestDTO;
+import com.themore.moamoatown.wish.dto.WishItemPurchaseResponseDTO;
+import com.themore.moamoatown.wish.dto.WishItemResponseDTO;
 import com.themore.moamoatown.wish.service.WishService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
 import java.util.List;
 
 /**
@@ -25,10 +26,7 @@ import java.util.List;
  * 2024.08.25  	임재성        최초 생성
  * 2024.08.25   임재성        위시 상품 조회 기능 추가
  * 2024.08.25   임재성        위시 상품 구매 기능 추가
- * 2024.08.26   임원정        위시 상품 생성, 삭제 추가
  * 2024.08.26   임재성        위시 상품 구매 메소드 수정
- * 2024.08.26   임원정        멤버 위시 상품 완료 처리 추가
- * 2024.08.27   임원정        멤버 위시 요청 리스트 조회 추가
  * </pre>
  */
 @Auth(role = Auth.Role.CITIZEN)
@@ -81,56 +79,6 @@ public class WishController {
 
         log.info("구매 완료 - " + response.getMessage());
 
-        return ResponseEntity.ok(response);
-    }
-
-    /**
-     * 위시 상품 생성
-     * @param requestDTO
-     * @param townId
-     * @return
-     */
-    @Auth(role = Auth.Role.MAYER)
-    @PostMapping("/create")
-    public ResponseEntity<String> createWishItem(@Valid @RequestBody WishItemCreateRequestDTO requestDTO,
-                                                 @TownId Long townId) {
-        wishService.createWishItem(requestDTO, townId);
-        return ResponseEntity.ok("위시 상품 생성이 완료 되었습니다.");
-    }
-
-    /**
-     * 위시 상품 삭제
-     * @param wishId
-     * @return
-     */
-    @Auth(role = Auth.Role.MAYER)
-    @DeleteMapping("/delete/{wishId}")
-    public ResponseEntity<String> deleteWishItem(@PathVariable Long wishId) {
-        wishService.deleteWishItem(wishId);
-        return ResponseEntity.ok("위시 상품 삭제가 완료 되었습니다.");
-    }
-
-    /**
-     * 멤버 위시 상품 완료
-     * @param memberWishId
-     * @return
-     */
-    @Auth(role = Auth.Role.MAYER)
-    @PatchMapping("/complete/{memberWishId}")
-    public ResponseEntity<String> completeWishItem(@PathVariable Long memberWishId) {
-        wishService.completeMemberWishItem(memberWishId);
-        return ResponseEntity.ok("위시 상품 사용 완료 처리 되었습니다.");
-    }
-
-    /**
-     * 멤버 위시 요청 리스트 조회
-     * @param townId
-     * @return
-     */
-    @Auth(role = Auth.Role.MAYER)
-    @GetMapping("/requests")
-    public ResponseEntity<List<MemberWishRequestsResponseDTO>> getMemberWishRequests(@TownId Long townId) {
-        List<MemberWishRequestsResponseDTO> response = wishService.getMemberWishRequests(townId);
         return ResponseEntity.ok(response);
     }
 }
