+++ src/main/java/com/themore/moamoatown/wish/mapper/WishMapper.java
@@ -1,8 +1,6 @@
 package com.themore.moamoatown.wish.mapper;
 
 
-import com.themore.moamoatown.wish.dto.MemberWishRequestsResponseDTO;
-import com.themore.moamoatown.wish.dto.WishItemCreateRequestDTO;
 import com.themore.moamoatown.wish.dto.WishItemPurchaseInternalRequestDTO;
 import com.themore.moamoatown.wish.dto.WishItemResponseDTO;
 import org.apache.ibatis.annotations.Mapper;
@@ -23,10 +21,7 @@ import java.util.List;
  * 2024.08.25   임재성        위시 상품 조회
  * 2024.08.25   임재성        위시 상품 구매
  * 2024.08.26   임재성        위시 상품 조회 메서드 수정
- * 2024.08.26   임원정        insertWish, deleteWish, deleteMemberWish 메소드 추가
  * 2024.08.26   임재성        위시 상품 구매 메소드 수정
- * 2024.08.26   임원정        updateMemberWishCompleted 메소드 추가
- * 2024.08.27   임원정        selectWishRequestsByTownId 메소드 추가
  * </pre>
  */
 @Mapper
@@ -44,17 +39,4 @@ public interface WishMapper {
      * @param internalDTO
      */
     void purchaseWishProcedure(WishItemPurchaseInternalRequestDTO internalDTO);
-
-    
-    // 위시 상품 생성
-    int insertWish(WishItemCreateRequestDTO createRequestDTO);
-    
-    /** 위시 삭제 **/
-    int deleteMemberWish(Long wishId);  // 멤버 위시 삭제
-    int deleteWish(Long wishId);    // 위시 삭제
-
-    // 멤버 위시 상품 완료 처리
-    int updateMemberWishCompleted(Long memberWishId);
-    // 위시 상품 요청 현황
-    List<MemberWishRequestsResponseDTO> selectWishRequestsByTownId(Long townId);
 }
