+++ src/main/resources/com/themore/moamoatown/quest/mapper/QuestMapper.xml
@@ -13,9 +13,6 @@
 * ==========  ========     ===========================
 * 2024.08.26  이주현        최초 생성
 * 2024.08.26  이주현        퀘스트 수락 요청 기능 추가
-* 2024.08.27  임원정        insertQuest, selectQuestStatusListByTownId 메소드 추가
-* 2024.08.28  임원정        selectMemberQuestByQuestId, updateMemberQuestSelected 메소드 추가
-* 2024.08.28  임원정        callCompleteQuestProcedure, findMemberIdByMemberQuestId 메소드 추가
 * </pre>
 -->
 
@@ -51,64 +48,4 @@
         VALUES (MEMBER_QUEST_SEQ.NEXTVAL, #{memberId}, #{questId}, 'N', 'N')
         ]]>
     </insert>
-
-    <!-- 퀘스트 삽입 -->
-    <insert id="insertQuest">
-        <![CDATA[
-            INSERT INTO QUEST (QUEST_ID, town_id, title, description, reward, capacity, deadline)
-            VALUES (QUEST_SEQ.NEXTVAL, #{townId}, #{title}, #{description}, #{reward}, #{capacity}, TO_DATE(#{deadline}, 'YYYY-MM-DD'))
-        ]]>
-    </insert>
-
-    <!-- 퀘스트 현황 조회 -->
-    <select id="selectQuestStatusListByTownId" resultType="QuestStatusListResponseDTO">
-        <![CDATA[
-            SELECT q.quest_id, q.title, q.reward, TO_CHAR(q.deadline, 'YYYY.MM.DD') as deadline, COUNT(mq.member_quest_id) as request_cnt, COUNT(DECODE(mq.selectedyn, 'Y', 1, NULL)) as selected_cnt, q.capacity
-            FROM quest q
-            LEFT JOIN member_quest mq ON q.quest_id = mq.quest_id
-            WHERE q.town_id = #{townId}
-            GROUP BY q.quest_id, q.title, q.reward, TO_CHAR(q.deadline, 'YYYY.MM.DD'), q.capacity
-            ORDER BY deadline DESC
-        ]]>
-    </select>
-
-    <!-- 퀘스트 ID로 member_quest 찾기 -->
-    <select id="selectMemberQuestByQuestId" resultType="MemberQuestRequestsResponseDTO">
-        <![CDATA[
-            SELECT
-                mq.member_quest_id,
-                m.nickname,
-                CASE
-                    WHEN mq.endYN = 'Y' THEN 3  -- 퀘스트 완료: 3
-                    WHEN mq.selectedYN = 'Y' THEN 2  -- 퀘스트 수락: 2
-                    ELSE 1 -- 수락 대기 중: 1
-                    END AS status
-            FROM member_quest mq
-            LEFT JOIN member m ON mq.member_id = m.member_id
-            WHERE mq.quest_id = #{questId}
-        ]]>
-    </select>
-
-    <!-- 퀘스트 요청 승인 -->
-    <update id="updateMemberQuestSelected">
-        <![CDATA[
-            UPDATE member_quest
-            SET selectedyn = 'Y'
-            WHERE member_quest_id = #{memberQuestId}
-        ]]>
-    </update>
-
-    <!-- 퀘스트 완료 처리_프로시저 호출 -->
-    <select id="callCompleteQuestProcedure" statementType="CALLABLE">
-        {CALL complete_quest(#{memberQuestId, jdbcType=NUMERIC})}
-    </select>
-
-    <!-- 퀘스트 요청 또는 완료 시 회원 ID 조회 -->
-    <select id="findMemberIdByMemberQuestId" resultType="Long">
-        <![CDATA[
-            SELECT member_id
-            FROM member_quest
-            WHERE member_quest_id = #{memberQuestId}
-        ]]>
-    </select>
 </mapper>
